+++ source/GLG3Dcpp/RenderDevice.cpp
@@ -1364,7 +1364,7 @@ void RenderDevice::enableClip2D(const Rect2D& clip) {
     minStateChange();
     minGLStateChange();
     state.clip2D = clip;
-    glScissor(clip.x0(), getHeight() - (int)(clip.y0() + clip.height()), (int)clip.width(), (int)clip.height());
+    glScissor((int)clip.x0(), getHeight() - (int)(clip.y0() + clip.height()), (int)clip.width(), (int)clip.height());
 
     if (! state.useClip2D) {
         glEnable(GL_SCISSOR_TEST);
